convert estimates comparison to new explorer API
